@@ -1,15 +1,12 @@
-import 'package:flutter/material.dart';
-
 class Transaction {
   final String id;
   final String title;
   final double amount;
   final DateTime date;
 
-  Transaction({
-    @required this.id,
-    @required this.title,
-    @required this.amount,
-    @required this.date
-    });
+  Transaction(
+      {required this.id,
+      required this.title,
+      required this.amount,
+      required this.date});
 }
