@@ -45,8 +45,8 @@ class Chart extends StatelessWidget {
             return Flexible(
               fit: FlexFit.tight,
               child: ChartBar(
-                  data['day'],
-                  data['amount'],
+                  data['day'].toString(),
+                  double.parse(data['amount'].toString()),
                   (totalSpending == 0
                       ? 0.0
                       : ((data['amount'] as double) / totalSpending))),
